analog sensors: add potentiometer functionality and print to LCD screen
Add potentiometer and read the value of the potentiometer through the
LCD screen as well as a mechanical switch that prints to the LCD screen
"on" or "off".

The value read by the potentiometer is a 10 bit number that should range
from around 0 to close to 4096 when the knob is turned to its minimum
and maximum degrees.

Angela Sodemann: Robotics 1 U2 (Sensors and Vision) S2 (Analog Sensors)
P1 (Potentiometer and ADC)