background subtraction: add different algorithm to calculate difference
Calculate difference in foreground and background images using
background subtraction technique. This technique is useful when we don't
know the color of the object we want to find. It only works if the image
is very stationary and the object enters or exits the field of view.

We need to use two while loops to first get the initial grayscale image of
comparison with no object in it. Then using that as the point of
comparison, we can then calculate the difference of a second grayscale
image frame with the object inside to find the pixel location of the
center of its brightness.

background_subtraction.py file as the main body of work and the
quiz_1_2_4 file to check handwritten work on python to calculate the X
position.

Angela Sodemann: Robotics 1 U2 (Sensors and Vision) S4 (Image Subtraction) P2 (Background Subtraction)