camera coordinates: add code to convert from camera to base grid
Implement algorithm to take the cm/pixels code from the camera
coordinates, to convert it to the base grid frame coordinates for our
manipulator. This is so the code will be useful for our manipulator and
now we can measure the location of objects on the board relative to our
grid frame, NOT the camera frame.

Add extra file to answer quiz 1-2-5 question 8 and 9.

Angela Sodemann: Robotics 1 U2 (Sensors and Vision) S5 (Camera Coordinates) P2 (Coordinate Transformation)