P2 pulse width modulation: Implement PWM functionality
Add code in PSoC creator to get the DC motor to work with PWM
functionality and test the speed of the motor at different compare
values.

Use Excel to graph the voltage to mean relationship and plot the
trendline on the scatterplot to see the Motor Velocity Constant (the
slope of the y = mx+b line).

Angela Sodemann: Robotics 1 U3 (Motion Control) S1 (DC Motors Speed and Torque) P2 (Pulse-Width Modulation)