P2 rack and pinion: modify code to use rack and pinion
Hook up the DC motor with the rack and pinion on the last link of the
2 degree of freedom revolute-revolute manipulator.

Use two target_position variables so the rack and pinion can move
between two points every 5 seconds. Notice how the control is still
unstable and when the rack comes to a puse, the motor is still
jittering violently.

Angela Sodemann: Robotics 1 U3 (Motion Control) S2 (Open-Loop vs Feedback) P2 (Rack and Pinion)