P1 Stability On-Off Control: Increase stability through tradeoffs
By decreasing the accuracy of the acceptable target value range of the
rack as well as reducing the speed of the motor, we are able to achieve
higher stability on the second pass of the rack and pinion movement
(second while loop).

To reduce the motor speed reduce the compare value in the
pwm_1_WriteCompare() functions.

To reduce the target accuracy, decrease and increase the minimum and
maximum range of the target_count2 variable in the if-else branches of
the second while loop.

Angela Sodemann: Robotics 1 U3 (Motion Control) S3 (On-Off vs Linear) P1 (Stabilizing On-Off Control)