P2 proportional control: fix instability in the rack and pinion movement
Modify the code to use proportional control instead of on and off
control. On and off control results in constantly overshooting the
target forwards and backwards which results in the motor violently
oscillating indefinitely. Proportional control implementation slows down
the speed of the motor proportional to the distance left to the target
position. This results in more stability.

The equation used to calculate the speed is the following:
speed = Kp * error

Where Kp is the control "gain" and error is the difference in the target
position to the current position. This is represented by the variable
"counter" in the code.

Angela Sodemann: Robotics 1 U3 (Motion Control) S3 (On-Off vs Linear) P2 (Proportional Control)