P1 computer to PSoC: Implement UART functionality from PC to PSoC
Use the python library "pyserial" to to set up the baudrate and open the
corresponding port to send integers from the range [0:255] to the LCD
display.

The UART block in the top design is set to 9600 baudrate and in the
main.c file we use the polling method to receive the byte sent from
Python.

Angela Sodemann: Robotics 1 U3 (Motion Control) S4 (UART) P1 (Computer to PSoC)