P2 PSoC to computer: Modify code to send encoder values from PSoC to PC
Use bit manipulation algorithm to send the counter values from 0 to 2000
as a set of 2, 8 bit bytes to avoid the limitation of only being able to
send up to the value 255 due to 8 bit restriction. Separate the two
bytes in the PSoC code and then reassemble it and save those values in
the Python code.

From the python code, use data.txt to copy the values and plot a
scatterplot in excel to see the position graph of the rack and pinion
for two movements.

Angela Sodemann: Robotics 1 U3 (Motion Control) S4 (UART) P2 (PSoC to Computer)