P1 time (step) response: plot graphs to see time response curve
Change the Kp value from [1.0:3.0] and collect the data of the encoder
counter for each Kp and transfer it to the excel sheet. Plot the 'time
response' graphs to see the relationship of how the joint responds to
our control over time.

We are interested in the tradeoffs between the speed, stability, and
accuracy of the control. Create a table to hold the measures of these
three categories. There will be 5 characteristics in the table.

Angela Sodemann: Robotics 1 U3 (Motion Control) S5 (PID) P1 (Time (Step) Response)