feat(servos): implement servo movement using input angles
Use a PWM block in PSoC Creator's top design to move the two revolute
joints to desired positions using angles as the input. The angles are
then calculated to the necessary compare values to write to the PWM
block for each servo.

The algebraic equation used is the 2 point form of the line equation
which is the following:

y - y1 = ((y2 - y1) / (x2 - x1)) * (x - x1)

where 'y' is the compare value, and 'x' is the angle on the linear
graph.

The min/max compare values were tuned to be as accurate as possible at
each 45.0 degree increment of angles. 0, 45, 90, 135, and 180 degrees.