feat (encoder): Add functionality to count value of dc motor encoder
Test to make sure the dc motor encoder is counting the correct
direction. Initially the quadrature decoder block in the top design had
the C1 as input A and C2 as input B but this resulted in the counter
counting backwards.

Instead we switched the C2 to be input A so this
way, when the prismatic joint drops down towards the board, the range is
from approximately 0 to 3000.

Also add functionality to print the counter on the LCD display to sanity
check the values prior to mounting the rack and pinion as a prismatic
joint.