feat (rack and pinion): Implement raising/lowering of rack and pinion
The two files 'motor.h' and 'motor.c' holds the implementation to raise
and lower the rack and pinion. The proportional control algorithm is
used to achieve this control instead of the "on-off" motion control to
achieve better stability and to remove the constant jittering and
oscillation of the motor to avoid damage.

TODO: Test the rack and pinion position in conjunction with the servos
position sequentially.