feat(servos): Modify code to move each servo one position at a time
The servos and rack and pinion are confirmed to work in unison and the
user can input desired angles and delay to move the servos independently
as well as lower and raise the rack and pinion.

Refactor manipulator_joint_position() functions as a function that moves
the servos only one position at a time and to also take in a time delay
parameter. This gives better control over the servos and to use these
functions in conjunction with the rack and pinion functionality.

Modify name of "scara_manipulator.*" files to "servos.*" file name for
better clarity of the function of the file.

Also rename the "pwm_1" top design block to "pwm_servos" to clarity the
functionality.

Change the 'raise' target value for the rack and pinion to be '0' for
easy testing purposes.