feat (xy positioning): refactor code to better implement encapsulation
Improve code so the relevant kinematics equations are encapsulated into
the "servos.c" file and "main.c" is not polluted with algorithmic code
and lots of variables.

The user can now just directly call the function
"servos_position_set_xy()" and input a desired X and Y coordinate to
move the manipulator, rather than modifying the X and Y variables in the
main function.

The links "a2" and "a4" were further tested and tuned, as well as
checking the workspace of the manipulator through brute force method
with a test function in the "test.c" file.