feat (image subtraction): Implement background image subtraction
Use machine vision techniques in order to capture the target object as
the brightest location on the camera's FOV. We use background
subtraction because we don't necessarily know the color of the object on
the RGB scale that we're trying to pick up (its a small metal nut).

This will be useful later so we can pinpoint the center of the mass of
the object using mathematical algorithms to calculate its center.