feat (center of mass): find location of the object in units of pixels
Use the 'center of mass' approach to get the object's location on the
board grid as a coordinate of 2 pixels for both rows and columns.

The user can now press the 'ESC' key on the initial 'background' image
to grab a snapshot of a clear board, then put the nut on the board and
it'll show up as a bright object in the 'difference' image frame. The
python shell will show the coordinates of the nut in pixels.

Also added a project notes section on the bottom to highlight critical
errors and debugging.