feat (pxls to cms): convert location from pixels to centimeters
Take measurement of the camera FOV using a ruler and get the dimensions
in centimeters. We know the cameras resolution is 640x480 where 640 is
the pixels in columns and 480 is the pixels for rows.

Calculate a ratio by taking the (camera fov in centimeters / in pixels)
and then multiply the row and column locations by that ratio to get the
units in centimeters.

Note the camera's base position is in the top left of the frame, but
this is not practical for our board since the camera is sitting
perpendicular to our base frame.