feat (coordinate transformation): convert from camera to base frame
Convert the coordinates of the object on the grid from the camera
coordinate system to the base frame coordinate system. The camera sits
perpendicular to the manipulator so using the coordinates from the
camera's POV doesn't make it practical to use in conjunction with our
manipulator.

All the equations and calculations are commented into the python code.
The accuracy of the object location is pretty accurate within ~0.1 - 0.2
cms which should be sufficient for our project.