feat (uart): Implement python and PSoC code to work together
Add PSoC and Python code so PSoC can receive the coordinates directly
from the Python's camera capture and then use it to move the servos
without hard coding coordinates. This is done through the "uart.c"
module. Add print_coordinates() function in this module as well to
confirm the received coordinates from python are the correct ones.

Add multiple test functions in order to debug the UART functionality
properly. Lots of testing was done to confirm UART worked to receive
both X and Y coordinates. These were done in a superloop to avoid having
the servos and motor move around and to test the functions in isolation
from the rest of the main.c code.

A lot of trial and error testing to make sure the servos and rack and
pinion would move in conjunction together to clear both the drop off bin
and the camera to avoid hitting them.

Refactor code in servos.c to add delay parameters to
servos_position_reset() function to better control the timing in unison
with the rack and pinion. Also add servos_position_before_dropoff()
function to allow the rack and pinion to raise high enough before moving
above the dropoff bin. This is to avoid hitting the bin away.

Refactor code in motor.h to add 'int target' parameters to both the
raise and lower functions. This is again, to allow more precise control
and more positional adjustments to avoid collisions between the
manipulator, the camera, and the dropoff bin.

Define macros at the top of main.c were added to better show intent of
the use of specific target parameters for raising and lowering the rack
and pinion.

TODO: fix Y position coordinate on LCD screen when its a double digit
number case