fix (offset): Add offset to x coordinate
The plastic bracket and the rack for the end effector flexes and bends
in a way where, the location of the magnet isn't exactly straight when
looking at it from the top down.

The added line "received_x -= 0.5" in the main.c code ensures that
received coordinates has the proper offset to account for this flexing
and bending of the plastic.

Also adjusted the values for a2 and a4 links in "servos.c" to better tune the
end effector.